new foc

+++ firmware_1/LibsCommon/gonio.h
@@ -3,15 +3,15 @@
 
 #include <stdint.h>
 
-#define SINE_TABLE_SIZE ((unsigned int)1024)
-#define SINE_TABLE_MAX  ((int16_t)512)
+#define SINE_TABLE_SIZE ((uint32_t)1024)
+#define SINE_TABLE_MAX  ((int32_t)512)
 
 
 //angle in range 0..SINE_TABLE_SIZE-1, unsigned
 //SINE_TABLE_MAX corresponds to 1 value
 //-SINE_TABLE_MAX corresponds to -1 value
 
-int16_t sin_tab(uint32_t angle);
-int16_t cos_tab(uint32_t angle);
+int32_t sin_tab(uint32_t angle);
+int32_t cos_tab(uint32_t angle);
  
 #endif
